requirements.txt
Django==5.2.8
tweepy==4.15.0
xdk==0.2.0b0
python-dotenv==1.0.0
psycopg2-binary==2.9.11
APScheduler==3.11.1